Added build system for firmware. Added some VSCode settings. Reorganised project for build.
