+++ code/application/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(OpenVent_application)
+
+add_subdirectory(flow)
