+++ code/application/flow/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_application
+  PUBLIC
+    flow.c
+)
