+++ code/drivers/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(OpenVent_drivers)
+
+add_subdirectory(???)
