+++ code/hal/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_library(OpenVent_hal)
+
+add_subdirectory(adc)
+add_subdirectory(board)
+add_subdirectory(clock)
+add_subdirectory(gpio)
+# add_subdirectory(spi)
+add_subdirectory(types)
+# add_subdirectory(uart)
