+++ code/hal/adc/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_hal
+  PUBLIC
+    adc_${MCU_TARGET}.c
+)
