+++ code/hal/clock/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_hal
+  PUBLIC
+    clock_${MCU_TARGET}.c
+)
