+++ code/hal/gpio/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_hal
+  PUBLIC
+    gpio_${MCU_TARGET}.c
+)
