+++ code/hal/spi/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_hal
+  PUBLIC
+    spi_${MCU_TARGET}.c
+)
