+++ code/hal/types/types.h
@@ -2,7 +2,7 @@
 #define TYPES_H
 
 #ifdef BOARD_MK1
-#include "mcu/types_avr.h"
+#include "types/types_avr.h"
 #endif
 
 #endif /* TYPES_H */
