+++ code/hal/uart/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(OpenVent_hal
+  PUBLIC
+    uart_${MCU_TARGET}.c
+)
